requirements.txt
streamlit==1.32.0
langchain==0.1.9
langchain-text-splitters==0.0.1
openai==1.14.0
python-dotenv==1.0.0
requests==2.31.0
pyaudio==0.2.14
ffmpeg-python==0.2.0
pydub==0.25.1
numpy==1.26.3
watchdog==3.0.0
